Add generic homebrew path finding option to work with custom homebrew setups and rely less on hardcoded variants
Should work for most homebrew setups.
Tested on a generic homebrew install on macos and a custom install on a Codam iMac (not tested with homebrew42 setup)
So this could, in theory, make the other two blocks obsolete, but I couldn't test with homebrew42, so that should be checked before.

Signed-off-by: Vincent@Codam <>