Run compile_shader.bat using unix file path
Windows cant find the ".\tools\compile_shader.bat" in the CI/CD.
The path it reports is ".toolscompile_shader.bat" which leads me to
believe it might expect a unix file path instead so we'll try that.
This is a long shot, but I'm losing my mind at window's behavior.