Run windows CI/CD with unix make file
I cant seem to figure out github action's windows.
The issue revolts around the shader compile batch script, no matter what
I do it just doesn't seem to function. With the current implantation it
attempts to run it with Bash and wants a unix file path. When specifying
that it should run using cmd with `cmd /C` it just doesn't function, it
seems like make just skips the rule. But thats not the case as when
printing debugging info it finds the rule and even runs it but nothing
happens, not even the print on the line above is seen in the logs.

From my testing it just wants a unix path for both the script and shader
file, so it might as well just use the unix makefile instead. Locally
the windows makefile functions just fine.

If someone has an answer that'd be great, but I'm going crazy and
switching to the unix makefile should be an easy fix.