Makefile cleanup and fixes (#24)
* Move #!/bin/bash to top of file

* Cleanup .PHONY

* Use correct terminology for make rules

* Fix re rule for when using multi-threading

* Always override HEADERS variable

* Use simply expanded variables for most variables

* Fix typo in generic homebrew path libglfw location

* Preprocessing indent consistency

* Fix typo

* use intended windows blank line echo

* Centralize make file variables

* Ensure no duplicates in OBJS

* Add some temporary debug prints for failing make

* Revert "Fix typo"

This reverts commit 2b2b2feba29cacc4b20abe0ff2afbba3c1316b1c.

* Switch windows CI/CD from Cygwin to MinGW

* Use Makefile_WindowsNT for windows CI/CD

* Revert "Add some temporary debug prints for failing make"

This reverts commit 4f79b31751400689b4dd943c63b20572e0572425.

* Remove NOWARNING flag

* Use powershell when running make in windows CI/CD

* CI/CD file cleanup

* Revert "Use powershell when running make in windows CI/CD"

This reverts commit e0e169daa3372359df64ea1cea741fe2b868625a.

* Explicitly run batch file in CMD from makefile

* Fix CI/CD workflow syntax error

* Guarantee unix file path from rwildcard function

* Add temporary debug printing to makefile

* Run make with debuging prints for windows CI/CD

* Move shader compile rule above C file compilation

This shouldn't have any impact, but honestly I might as well try.
Windows just be weird and funky.

* Revert "Add temporary debug printing to makefile"

This reverts commit 9fb1ec83950bcccf129de9d39553ce046544a7ca.

* Revert "Run make with debuging prints for windows CI/CD"

This reverts commit 3706ce05093beefa6b3d82bcc39f8c7ada2b5997.

* Revert "Move shader compile rule above C file compilation"

This reverts commit 0d6ba2a61244958f84bfb6172a97bb62d997b879.

* Revert "Explicitly run batch file in CMD from makefile"

This reverts commit 138c6442b3a9efb9afc1fc9ba07cbe2bc1b29df6.

* Run compile_shader.bat using unix file path

Windows cant find the ".\tools\compile_shader.bat" in the CI/CD.
The path it reports is ".toolscompile_shader.bat" which leads me to
believe it might expect a unix file path instead so we'll try that.
This is a long shot, but I'm losing my mind at window's behavior.

* Revert "Run compile_shader.bat using unix file path"

This reverts commit 3e5c0520066ff78cbedf677a3b179568f170746e.

* Run windows CI/CD with unix make file

I cant seem to figure out github action's windows.
The issue revolts around the shader compile batch script, no matter what
I do it just doesn't seem to function. With the current implantation it
attempts to run it with Bash and wants a unix file path. When specifying
that it should run using cmd with `cmd /C` it just doesn't function, it
seems like make just skips the rule. But thats not the case as when
printing debugging info it finds the rule and even runs it but nothing
happens, not even the print on the line above is seen in the logs.

From my testing it just wants a unix path for both the script and shader
file, so it might as well just use the unix makefile instead. Locally
the windows makefile functions just fine.

If someone has an answer that'd be great, but I'm going crazy and
switching to the unix makefile should be an easy fix.

* Add cross platform mlx_getline

* Small description patch up

* Rename CYGWIN to WIN_UNIX

* Move * to type not name

* Typo fix

* Move GETLINE_BUFF define to header

* Remove NDEBUG todo

* Add missing override in front of HEADERS

* Fix make error has to have no identation

* Reuse the same memory in mlx_getline

* Apply some adjustments

Co-authored-by: W2.Wizard <63303990+W2Wizard@users.noreply.github.com>